UI fixes + placeholder changes
