+++ src/components/CIPipelineN/CustomInputOutputVariables.tsx
@@ -169,41 +169,41 @@ function CustomInputOutputVariables({ type }: { type: PluginVariableType }) {
     return (
         <>
             <div className="row-container mb-4 mt-4">
-                <Tippy
-                    className="default-tt"
-                    arrow={false}
-                    content={
-                        <span style={{ display: 'block', width: '220px' }}>
-                            {type === PluginVariableType.INPUT
-                                ? TaskFieldDescription.INPUT
-                                : TaskFieldDescription.OUTPUT}
-                        </span>
-                    }
-                >
-                    <div className={`tp-4 fs-13 fw-6 mr-8 `} style={{ position: 'relative' }}>
+                <div className={`tp-4 fs-13 fw-6 mr-8 `} style={{ position: 'relative' }}>
+                    <Tippy
+                        className="default-tt"
+                        arrow={false}
+                        content={
+                            <span style={{ display: 'block', width: '220px' }}>
+                                {type === PluginVariableType.INPUT
+                                    ? TaskFieldDescription.INPUT
+                                    : TaskFieldDescription.OUTPUT}
+                            </span>
+                        }
+                    >
                         <span className="text-underline-dashed lh-32">{type} variables </span>
-                        {isDateFormat && (
-                            <div
-                                className="bcb-1 br-4 fw-4 pl-12 pr-12 pt-8 pb-8"
-                                style={{ width: '180px', marginRight: '60px', position: 'absolute' }}
-                            >
-                                <div className="format-grid">
-                                    <Info className="mr-4 icon-dim-16" />
-                                    <span className="cb-5 mb-2 lh-1-33">
-                                        <a
-                                            className="no-decor"
-                                            href="https://github.com/Knetic/govaluate/blob/0580e9b47a69125afa0e4ebd1cf93c49eb5a43ec/parsing.go#L258"
-                                            target="_blank"
-                                        >
-                                            Standardized date formats
-                                        </a>{' '}
-                                        <span className="cn-9">identified by Devtron</span>{' '}
-                                    </span>
-                                </div>
+                    </Tippy>
+                    {isDateFormat && (
+                        <div
+                            className="bcb-1 br-4 fw-4 pl-12 pr-12 pt-8 pb-8"
+                            style={{ width: '180px', marginRight: '60px', position: 'absolute' }}
+                        >
+                            <div className="format-grid">
+                                <Info className="mr-4 icon-dim-16" />
+                                <span className="cb-5 mb-2 lh-1-33">
+                                    <a
+                                        className="no-decor"
+                                        href="https://github.com/Knetic/govaluate/blob/0580e9b47a69125afa0e4ebd1cf93c49eb5a43ec/parsing.go#L258"
+                                        target="_blank"
+                                    >
+                                        Standardized date formats
+                                    </a>{' '}
+                                    <span className="cn-9">identified by Devtron</span>{' '}
+                                </span>
                             </div>
-                        )}
-                    </div>
-                </Tippy>
+                        </div>
+                    )}
+                </div>
 
                 <div className="pointer cb-5 fw-6 fs-13 flexbox content-fit lh-32" onClick={addVariable}>
                     <Add className="add-icon mt-6" />
