+++ src/components/CIPipelineN/TaskFieldTippyDescription.tsx
@@ -9,24 +9,22 @@ interface TippyDescriptionType {
 
 function TaskFieldTippyDescription({ taskField, contentDescription }: TippyDescriptionType) {
     return (
-        <div>
+        <div className="fw-6 fs-13 lh-32 cn-7" style={{ maxWidth: '220px' }}>
             <Tippy
                 className="default-tt"
                 arrow={false}
                 content={<span style={{ display: 'block', width: '220px' }}>{contentDescription}</span>}
             >
-                <div className="fw-6 fs-13 lh-32 cn-7" style={{ maxWidth: '220px' }}>
-                    <span className="text-underline-dashed">
-                        {taskField}
-                        {taskField === TaskFieldLabel.SCRIPT ||
-                        taskField === TaskFieldLabel.MOUNTCODEAT ||
-                        taskField === TaskFieldLabel.CONTAINERIMAGEPATH ? (
-                            <span className="cr-5"> *</span>
-                        ) : (
-                            ''
-                        )}
-                    </span>
-                </div>
+                <span className="text-underline-dashed">
+                    {taskField}
+                    {taskField === TaskFieldLabel.SCRIPT ||
+                    taskField === TaskFieldLabel.MOUNTCODEAT ||
+                    taskField === TaskFieldLabel.CONTAINERIMAGEPATH ? (
+                        <span className="cr-5"> *</span>
+                    ) : (
+                        ''
+                    )}
+                </span>
             </Tippy>
         </div>
     )
